feat : Comment update,delete 유저인중 추가

+++ src/main/kotlin/com/b05/newsfeedproject/domain/comments/service/CommentServiceImpl.kt
@@ -5,6 +5,7 @@ import com.b05.newsfeedproject.domain.comments.dto.CreateCommentRequest
 import com.b05.newsfeedproject.domain.comments.dto.UpdateCommentRequest
 import com.b05.newsfeedproject.domain.comments.model.Comment
 import com.b05.newsfeedproject.domain.comments.repsitory.CommentRepository
+import com.b05.newsfeedproject.domain.exception.AccessDeniedException
 import com.b05.newsfeedproject.domain.exception.ModelNotFoundException
 import com.b05.newsfeedproject.domain.posts.repository.PostRepository
 import com.b05.newsfeedproject.domain.user.repository.UserRepository
@@ -47,6 +48,8 @@ class CommentServiceImpl(
         commentId: Int,
         request: UpdateCommentRequest
     ): CommentResponse {
+        val user = userRepository.findByIdOrNull(userId)
+            ?: throw ModelNotFoundException("user", userId)
         val post = postRepository.findByIdOrNull(postId)
             ?: throw ModelNotFoundException("post", postId)
         val comment = commentRepository.findByIdOrNull(commentId)
@@ -54,17 +57,25 @@ class CommentServiceImpl(
         if (comment.post.id != post.id) {
             throw IllegalStateException("Post ID does not match with the comment's post ID")
         }
+        if (comment.user.id != user.id) {
+             throw AccessDeniedException("comment")
+        }
         comment.content = request.content
         return commentRepository.save(comment).toResponse()
     }
 
     override fun deleteComment(userId: Int, postId: Int, commentId: Int) {
+        val user = userRepository.findByIdOrNull(userId)
+            ?: throw ModelNotFoundException("user", userId)
         val post = postRepository.findByIdOrNull(postId)
             ?: throw ModelNotFoundException("post", postId)
         val comment = commentRepository.findByIdOrNull(commentId)
             ?: throw ModelNotFoundException("comment", commentId)
         if (comment.post.id != post.id) {
             throw IllegalStateException("Post ID does not match with the comment's post ID")
+        }
+        if (comment.user.id != user.id) {
+             throw AccessDeniedException("comment")
         }
         commentRepository.delete(comment)
     }
